Pull recipes from api
Created a RecipeService class to encapsulate any calls we have to the spoonacular api. Then did some work in the DiscoverPage.xaml.cs class--created a method (LoadRecipesAsync) which makes use of the RecipeService class to read recipes from the api then store them in the Recipes variable in the discover page.